Use byte-sized encoder constructor arguments

+++ src/mcl/MCL/MCLEncoder.h
@@ -22,7 +22,7 @@ class MCLEncoder : public Encoder {
        name, initial value, and handling function. The initRangeEncoder
        will be called with the constructor arguments.
      **/
-    MCLEncoder(int _max = 127, int _min = 0, int res = 1, int _speed = 4) : Encoder() {
+    MCLEncoder(uint8_t _max = 127, uint8_t _min = 0, uint8_t res = 1, uint8_t _speed = 4) : Encoder() {
       initMCLEncoder(_max, _min, (int) 0, res, (encoder_handle_t) nullptr);
       fast_speed = _speed;
     }
@@ -63,13 +63,13 @@ class MCLEncoder : public Encoder {
 
 class MCLExpEncoder : public MCLEncoder {
   public:
-  MCLExpEncoder(int _max = 127, int _min = 0, int res = 1) : MCLEncoder(_max,_min,res) { }
+  MCLExpEncoder(uint8_t _max = 127, uint8_t _min = 0, uint8_t res = 1) : MCLEncoder(_max,_min,res) { }
   virtual int update(encoder_t *enc);
 };
 
 class MCLRelativeEncoder : public MCLEncoder {
   public:
-  MCLRelativeEncoder(int _max = 127, int _min = 0, int res = 1) : MCLEncoder(_max,_min,res) { }
+  MCLRelativeEncoder(uint8_t _max = 127, uint8_t _min = 0, uint8_t res = 1) : MCLEncoder(_max,_min,res) { }
   virtual int update(encoder_t *enc);
 };
